Fixed security bugs for view orders/products,delete orders,users,products etc...

+++ backend/deleteMessages.php
@@ -1,9 +1,16 @@
 <?php
 include_once 'mappers/messagesMapper.php';
-if (isset($_GET['id'])) {
+session_start();
+if($_SESSION['role']==0){
+    header("Location:../index.php");
+}
+
+else{
+    if (isset($_GET['id'])) {
     $userId = $_GET['id'];
     $mapper = new MessagesMapper();
     $mapper->deleteMessages($userId);
     header("Location:../pages/Dashboard.php?id=messages");
-}
+    }   
 
+}
